Further split CLI TUI controller mixins
